third_party/WebKit/LayoutTests/http/tests/inspector/sources/inline-module-export-error-expected.txt

CONSOLE ERROR: line 5: Uncaught SyntaxError: The requested module does not provide an export named 'x'
CONSOLE ERROR: line 5: Uncaught SyntaxError: The requested module does not provide an export named 'x'
Tests inline module import error.

Script execution paused.
line 4, content: <script type="module">
Script execution resumed.
